Change nginx configurations
Default nginx configurations are now being stored and controlled in a
single place under version control. Also, I've changed configuration so
that the root of the default nginx site is pointing to a shared vagrant
directory. With this I can removed commands that have created a sym-link
of index pages from vagrant directory to a nginx-specific directory.